backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
openai==1.54.0
playwright==1.48.0
python-multipart==0.0.12
pydantic==2.9.0
python-dotenv==1.0.1
nest-asyncio==1.6.0